+++ test/unit/extract.js
@@ -13,7 +13,7 @@ describe('Partial image extraction', function () {
         if (err) throw err;
         assert.strictEqual(20, info.width);
         assert.strictEqual(20, info.height);
-        fixtures.assertSimilar(fixtures.expected('extract.jpg'), data, { threshold: 8 }, done);
+        fixtures.assertSimilar(fixtures.expected('extract.jpg'), data, done);
       });
   });