+++ test/unit/resize.js
@@ -386,33 +386,6 @@ describe('Resize dimensions', function () {
     });
   });
 
-  it('Centre vs corner convention return different results', function (done) {
-    sharp(fixtures.inputJpg)
-      .resize(32, 24, { centreSampling: false })
-      .greyscale()
-      .raw()
-      .toBuffer(function (err, cornerData) {
-        if (err) throw err;
-        assert.strictEqual(768, cornerData.length);
-        sharp(fixtures.inputJpg)
-          .resize(32, 24, { centerSampling: true })
-          .greyscale()
-          .raw()
-          .toBuffer(function (err, centreData) {
-            if (err) throw err;
-            assert.strictEqual(768, centreData.length);
-            assert.notStrictEqual(0, cornerData.compare(centreData));
-            done();
-          });
-      });
-  });
-
-  it('Invalid centreSampling option', function () {
-    assert.throws(function () {
-      sharp().resize(32, 24, { centreSampling: 1 });
-    });
-  });
-
   it('Dimensions that result in differing even shrinks on each axis', function (done) {
     sharp(fixtures.inputJpg)
       .resize(645, 399)
@@ -451,25 +424,19 @@ describe('Resize dimensions', function () {
 
   it('fastShrinkOnLoad: false ensures image is not shifted', function (done) {
     return sharp(fixtures.inputJpgCenteredImage)
-      .resize(9, 8, {
-        fastShrinkOnLoad: false,
-        centreSampling: true
-      })
+      .resize(9, 8, { fastShrinkOnLoad: false })
       .png()
       .toBuffer(function (err, data, info) {
         if (err) throw err;
         assert.strictEqual(9, info.width);
         assert.strictEqual(8, info.height);
-        // higher threshold makes it pass for both jpeg and jpeg-turbo libs
-        fixtures.assertSimilar(fixtures.expected('fast-shrink-on-load-false.png'), data, { threshold: 7 }, done);
+        fixtures.assertSimilar(fixtures.expected('fast-shrink-on-load-false.png'), data, done);
       });
   });
 
   it('fastShrinkOnLoad: true (default) might result in shifted image', function (done) {
     return sharp(fixtures.inputJpgCenteredImage)
-    .resize(9, 8, {
-      centreSampling: true
-    })
+    .resize(9, 8)
     .png()
     .toBuffer(function (err, data, info) {
       if (err) throw err;